+++ attest/core/src/lib.rs
@@ -27,9 +27,7 @@ pub use crate::{
         QuoteVerifyError, ReportBodyVerifyError, ReportDetailsError, RevocationCause, SgxError,
         SgxResult, SignatureError, TargetInfoError, VerifyError,
     },
-    ias::verify::{
-        EpidPseudonym, VerificationReport, VerificationReportData, VerificationSignature,
-    },
+    ias::verify::{EpidPseudonym, VerificationReportData},
     nonce::{IasNonce, Nonce, QuoteNonce},
     quote::{Quote, QuoteSignType},
     report::Report,
@@ -55,6 +53,8 @@ pub use crate::{
         ConfigSecurityVersion, MiscSelect, ProductId, SecurityVersion,
     },
 };
+
+pub use mc_attest_verifier_types::{VerificationReport, VerificationSignature};
 
 /// The IAS version we support
 pub const IAS_VERSION: f64 = 4.0;
