+++ attest/core/src/quote.rs
@@ -17,7 +17,6 @@ use crate::{
 use alloc::vec::Vec;
 use core::{
     cmp::{max, min},
-    convert::{TryFrom, TryInto},
     fmt::{Debug, Display, Formatter, Result as FmtResult},
     ops::Range,
 };
